move boundary/volume shape classification to base, override STL to work with its quirks
